Realizar rol y productos

+++ ArtInk/ArtInk.Application/Services/Interfaces/IServiceProducto.cs
@@ -5,6 +5,6 @@ namespace ArtInk.Application.Services.Interfaces
     public interface IServiceProducto
     {
         Task<ICollection<ProductoDTO>> ListAsync();
-        Task<ProductoDTO> FindByIdAsync(int id);
+        Task<ProductoDTO> FindByIdAsync(short id);
     }
 }
